move crash email to logger
